Reduce memory usage when decoding
The decoder interface now uses callbacks to set the layer pixel data
and the image meta data.
This removes the need to store that information as part of the decoder
state.

The layer bitmap is allocated when JXL_DEC_NEED_IMAGE_OUT_BUFFER is
received, and the image data is then copied to the output layer as it
becomes available.
This change also allows the frame name to be uses as the layer name, if
one is present.