Add support for loading CMYK images
CMYK JPEG XL images are identified by the presence of an extra channel
containing the black/key color data.
JPEG XL uses 0 to represent black/full ink for CMYK images.

https://discord.com/channels/794206087879852103/804324493420920833/1317698217273458738
"The K channel of a CMYK image. If present, a CMYK ICC profile is also
present, and the RGB samples are to be interpreted as CMY, where 0
denotes full ink."

As Windows Imaging Component uses 0 to represent white/no ink, we have
to invert the CMYK data before sending it to Paint.NET.

This commit depends on features added in Paint.NET 5.1.2, specifically
the ColorCmyk32 structure and the ability to create Cmyk32 images
using the Paint.NET WIC wrappers.