Load the first frame of a multi-frame image or animation
This is more user friendly than just refusing to read multi-frame
images or animations.

The main decoding loop has been split into two passes.
The first pass reads the first frame from the image, and the second
pass reads the EXIF and XMP metadata.
The second pass will only run if the JXL image has a container, as
one is required to store EXIF and/or XMP.