Load gray images as RGB
Gray images are loaded as RGB due to WIC having poor support for gray
to RGB format conversions.
WIC was throwing an exception when trying to convert from a gray color
profile to a RGB color profile.

The color profiles will always be RGB, even if the original image was
gray.
If the gray image has an ICC profile, it will be ignored.